refactor : Fix lint error - Remove unused import statements - Remove unused pages
