+++ .eslintrc.json
@@ -32,7 +32,11 @@
     "import/extensions":"off",
     "react/prop-types":"off",
     "@typescript-eslint/no-unsafe-member-access": "off",
-    "@typescript-eslint/no-unsafe-assignment":"off"
+    "@typescript-eslint/no-unsafe-assignment":"off",
+    "@typescript-eslint/no-unsafe-argument":"off",
+    "@typescript-eslint/no-unsafe-call": "off",
+    "@typescript-eslint/no-unsafe-return": "off",
+    "@typescript-eslint/restrict-template-expressions": "off"
   },
   "settings": {
     "react": {
