--- src/components/DropdownModal/index.ts
@@ -1,1 +1,0 @@
-export * from './DropdownModal';
