+++ src/pages/Landing/agreement/index.tsx
@@ -1,9 +1,8 @@
 import { ChangeEventHandler } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
-import { v4 as uuid } from 'uuid';
 import { InputContainer, StatusButton } from '@/components';
-import { agreementState, userState } from '@/state/atom';
+import { agreementState } from '@/state/atom';
 import { tw } from '@/utils/tailwindMerge';
 import { useCommutes } from '@/hooks/useCommutes';
