+++ src/routes/index.tsx
@@ -1,6 +1,5 @@
 import { Fragment } from 'react';
 import { Routes, Route } from 'react-router-dom';
-import { RecoilRoot } from 'recoil';
 
 interface Module {
   [modulePath: string]: { default: string };
@@ -36,7 +35,6 @@ export const Router = () => {
   const NotFound = preserved?.['404'] || Fragment;
 
   return (
-    <RecoilRoot>
       <App>
         <Routes>
           {routes.map(({ path, component: Component = Fragment }) => (
@@ -45,6 +43,5 @@ export const Router = () => {
           <Route path="*" element={<NotFound />} />
         </Routes>
       </App>
-    </RecoilRoot>
   );
 };
